+++ packages/menu/src/MenuItem.tsx
@@ -8,7 +8,7 @@ import { FocusRing } from '@react-aria/focus';
 import { cx } from 'classix';
 import { cloneElement } from 'react';
 
-import './styles/Menu.css';
+import styles from './styles/Menu.module.css';
 
 // Merge two types and get rid of overlapping definitions
 type Merge<T, U> = Omit<T, keyof U> & U;
@@ -76,17 +76,17 @@ const MenuItem = <P, T extends ElementType = typeof defaultElement>({
   const renderIcon = icon && cloneElement(icon, { size: 'small' });
 
   const renderedItem = (
-    <FocusRing focusRingClass="has-focus">
+    <FocusRing focusRingClass={styles['has-focus']}>
       <Component
         {...rest}
         disabled={disabled}
         aria-disabled={disabled ? disabled : undefined}
         className={cx(
-          'Menu-item',
+          styles['Menu-item'],
           className,
-          isHighlighted && 'is-highlighted',
-          nested && 'Menu-item--nested',
-          groupHeader && 'Menu-item--header'
+          isHighlighted && styles['is-highlighted'],
+          nested && styles['Menu-item--nested'],
+          groupHeader && styles['Menu-item--header']
         )}
         data-test-id={testId}
         role={role}
@@ -96,7 +96,7 @@ const MenuItem = <P, T extends ElementType = typeof defaultElement>({
           children
         ) : (
           <>
-            {icon && <span className="Menu-item-icon">{renderIcon}</span>}
+            {icon && <span className={styles['Menu-item-icon']}>{renderIcon}</span>}
             {children}
           </>
         )}