+++ packages/menu/src/MenuSearch.tsx
@@ -3,7 +3,7 @@ import type { ChangeEvent } from 'react';
 import { TextField } from '@launchpad-ui/form';
 import { forwardRef } from 'react';
 
-import './styles/Menu.css';
+import styles from './styles/Menu.module.css';
 
 type MenuSearchProps = {
   ariaLabel?: string;
@@ -24,11 +24,11 @@ const MenuSearch = forwardRef<HTMLInputElement, MenuSearchProps>((props, ref) =>
   } = props;
 
   return (
-    <div className="Menu-search">
+    <div className={styles['Menu-search']}>
       <TextField
         {...finalProps}
         ref={ref}
-        className="Menu-search-input"
+        className={styles['Menu-search-input']}
         tiny
         id={id}
         type="search"