+++ packages/tokens/stories/tokens.stories.tsx
@@ -1,6 +1,6 @@
 import { CopyToClipboard } from '@launchpad-ui/clipboard';
-
-import tokens from '../dist/default.json';
+import { vars } from '@launchpad-ui/vars';
+import { useEffect, useState, useRef } from 'react';
 
 export default {
   title: 'Tokens/Colors',
@@ -23,39 +23,54 @@ const flatten = (obj: Record<string, unknown>) => {
   return result;
 };
 
-const getTokenTable = (tokens: Record<string, unknown>) => (
-  <table style={{ borderCollapse: 'separate', borderSpacing: '20px 0' }}>
-    <thead>
-      <tr>
-        <th></th>
-        <th style={{ textAlign: 'left' }}>Name</th>
-        <th style={{ textAlign: 'left' }}>Value</th>
-      </tr>
-    </thead>
-    <tbody>
-      {Object.entries(tokens).map(([key, value]) => (
-        <tr key={key}>
-          <td>
-            <div
-              style={{
-                background: value,
-                height: '50px',
-                width: '150px',
-                border: '1px solid var(--lp-color-border-ui-primary)',
-              }}
-            ></div>
-          </td>
-          <td>
-            <CopyToClipboard text={`--lp-color-${key}`}>{`--lp-color-${key}`}</CopyToClipboard>
-          </td>
-          <td>{value}</td>
-        </tr>
-      ))}
-    </tbody>
-  </table>
-);
+const TokenTable = ({ tokens }: { tokens: Record<string, string> }) => {
+  const itemEls = useRef<Record<string, HTMLDivElement | null>>({});
+  const [colors, setColors] = useState<Record<string, string>>({});
 
-const global = Object.keys(tokens.color)
+  useEffect(() => {
+    for (const [key, value] of Object.entries(itemEls.current)) {
+      const item = { [key]: getComputedStyle(value as Element).backgroundColor };
+      setColors((c) => ({ ...c, ...item }));
+    }
+  }, []);
+
+  return (
+    <table style={{ borderCollapse: 'separate', borderSpacing: '20px 0' }}>
+      <thead>
+        <tr>
+          <th></th>
+          <th style={{ textAlign: 'left' }}>Name</th>
+          <th style={{ textAlign: 'left' }}>Value</th>
+        </tr>
+      </thead>
+      <tbody>
+        {Object.entries(tokens).map(([key, value]) => {
+          return (
+            <tr key={key}>
+              <td>
+                <div
+                  ref={(element) => (itemEls.current[key] = element)}
+                  style={{
+                    background: value,
+                    height: '50px',
+                    width: '150px',
+                    border: '1px solid var(--lp-color-border-ui-primary)',
+                  }}
+                ></div>
+              </td>
+              <td>
+                <CopyToClipboard text={`--lp-color-${key}`}>{`--lp-color-${key}`}</CopyToClipboard>
+              </td>
+              <td>{colors[key]}</td>
+            </tr>
+          );
+        })}
+      </tbody>
+    </table>
+  );
+};
+
+const global = Object.keys(vars.color)
   .filter((key) =>
     [
       'black',
@@ -71,21 +86,21 @@ const global = Object.keys(tokens.color)
     ].includes(key)
   )
   .reduce((obj, key) => {
-    obj[key] = tokens.color[key];
+    obj[key] = vars.color[key];
     return obj;
   }, {});
 
 export const Global = {
-  render: () => getTokenTable(flatten(global)),
+  render: () => <TokenTable tokens={flatten(global)} />,
 };
 
-const alias = Object.keys(tokens.color)
+const alias = Object.keys(vars.color)
   .filter((key) => ['bg', 'border', 'fill', 'shadow', 'text'].includes(key))
   .reduce((obj, key) => {
-    obj[key] = tokens.color[key];
+    obj[key] = vars.color[key];
     return obj;
   }, {});
 
 export const Alias = {
-  render: () => getTokenTable(flatten(alias)),
+  render: () => <TokenTable tokens={flatten(alias)} />,
 };